app/Http/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pinecone-client==3.0.0
openai==1.55.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-huggingface>=0.1.0
huggingface_hub>=0.23.0
pydantic>=2.0.0
